Use shared partial for confirmation screens
The confirmation screens for the edit metadata / edit facets forms
are almost identical, so we have an opportunity to DRY this up.

Crucially, the value passed to the `proposed_schema` hidden input
is absolutely vital - we'll cover this in the next commit or two.
Using a shared partial minimises the risk of us accidentally
breaking the value in one of the confirmation screens.

This was evidenced by the 'facets' confirmation page assigning
a value of `JSON.pretty_generate(@proposed_schema)` vs the
'metadata' confirmation page assigning a value of
`@proposed_schema.to_json`. The latter is the correct value. The
former is actually a critical bug - its output is something like
`<# FinderSchema::abcdef123456>`, meaning the payload information
is entirely lost.

This will be tested and tightened up in the following commit.