Pretty-print JSON, and test JSON submission value
The schema gets rendered to a hidden input as a FinderSchema instance
with `to_json` called on it, passed into `JSON.parse`, passed into
`JSON.pretty_generate`, which renders a multi-line string. That is
the value that is POST-ed to the controller.

This test started off life in the admin_controller_spec.rb, but
I realised we could pass any value to the `post :zendesk` call,
so it wasn't a real world test. The only way we can reliably test
this behaviour is to check the value of the hidden input itself,
coupled with the knowledge that the controller doesn't do any
transformation of the submitted value.

Pretty-printing the JSON should lead to a higher quality Zendesk
ticket being created, insofar as we can copy and paste the generated
schema into GitHub and see exactly what has changed in the diff.
Without prettification, the entire file will look like it's
changed.