Merge facet and sub-facet attributes for view rendering
When we load the edit page or new page again, the document is loaded from Publishing API, with the facet and sub-facet being separate facets, but the dropdown element expects a combined leaf node dropdown, so we have to merge them back together again into the main facet field.

We need to collate the values for the `ipo_grounds_section` and `ipo_grounds_sub_section` fields in the one field we do render (`ipo_grounds_section`).