Map single nested facet from schema to two individual facets for publishing-api
In order to minimise the customisation in Finder Frontend, we will send the data in the shape it is typically expected by Finder Frontend, i.e. the facets are stand-alone. We still need some way of referencing the parent facet value in the children, which we're doing by setting `main_facet_value`, alongside every subfacet label-value pair in the subfacet. We also send the `main_facet_label` so we can recompose the subfacet label in the format `Main facet label - Sub facet label` when JS is disabled and all the facet options show in the select; since we have duplicates such as "Not applicable" it would be unclear which main facet option the subfacet option refers to.

Other required keys are:
- the `nested_facet` (T/F) flag which indicates to the Finder Frontend rendering app that the facet should be rendered as a NestedFacet (new Facet type added by us).
- `sub_facet_key` and `sub_facet_name` which are only set on the main facet, used for loading the js module.
- the `key` and `name` of the subfacet come from the data stored under the main facet, in the schema.
- all the other key values are shared between the main and sub facets, though a case could be made in the future for also allowing the customization of the subfacet's `short_name`, `preposition`, `display_as_result_metadata` or `show_option_select_filter` values.

Co-authored-by: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>