Display subfacets in the form
This commit enables the rendering of subfacet values from the schema.

In the schema, a nested facet includes a `sub_facets` key with additional label-value pairs. As per this commit, the main facet's values and their subfacet's values are mapped to "blocks" of facets:
 - the first line is the main facet (which will slow as label only on a new form "Main Facet" and label-key on edit form "Main Facet{main-facet}")
 - subsequent rows include the subfacets in the same style: "Sub Facet", and "Sub Facet{sub-facet}" respectively.
 - each facet block is separated from the next by an empty line.

The splitting of the allowed value text area string (see subsequent commits) required that I render the two new lines as "\r\n\r\n" - it basically means one **visible** blank line.