Generate nested facet schema
We permit a new controller params (`sub_facet`) for passing data required for a nested facet processing. We require the `name` and `key` of a sub-facet, which are here mixed in this single input called `sub-facet`. This value will be given as "Facet {key}", as we do elsewhere on the form.

The processing in the `Facet` model will output:
- `nested_facet` - a boolean which we deduce from the `sub-facet` input being present
- sub-facet `key`
- sub-facet `name`
- the `allowed_values` of the facet, which is a nested structure containing the labels and values of the sub-facet, mixed in with the corresponding labels and values of the main facet.