Truncate value string to silence code scanning alert
There was a `Polynomial regular expression used on uncontrolled data` alert due to the potential length of the string on which we're then calling match.

The fix is to truncate the string. The assumption here is that any such string would include a very long key, so the cropping includes an ending "}".