Add a general test for creating new documents
Have consolidated a lof of the 'create an x' test logic into one general test file.

 The test suite includes the following scenarios:
 - Getting to the new document page.
 - Creating a document with valid data.
 - Attempting to create a document with no data.
 - Attempting to create a document with invalid content.
 - Attempting to create a document with an invalid date.

There are scenarios that have been removed in said consolidation, such as those around
the 'open before closed date' validation testing or test around permissions.

In the case of the former this has been moved to a seperate test file, all the other scenarios
I judged there to be suffcient coverage in existing tests (eg access_control).

The test suite dynamically generates tests for each document schema found in the "lib/documents/schemas" directory,
except for those listed in the EXCEPTIONS_TO_GENERAL_TESTING constant.

The exceptions constant is mostly there to account for the 5 finders that do not yet used the shared form.
The Marine equipment and protect food finders have custom validation on some of their fields, I considered accounting
for that in this test but ultimately did not do that as it started to stray away from a 'general test'. Therefore I have
exempted them from this general testing.

The base paths for the product recall and the ai assurance finders differ from the rest of the finders in that they are not just the
format type pluralised. I have accounted for this in the decribe block, it seemed to me to be the least intrusive way of handling the issue.