Add validation to remaining document type models
Given that we have adopted the approach of making model validation schema driven we
should make sure these validations are actually being applied on the models.

They were not added previously to models that did not have any validation listed in their schemas.
We have however decided that date validation should be applied universally based on the
type of the facet in the schema, so it it necessary to enforce this in the models.