Consolidate model spec in general test file
Our validations are now largely defined by the schemas and are being applied
to the models without the need to be manually defined and configured.

This validation in tested by our feature tests. It seems unnecessary to test this validation twice.

The ai_assurance_portfolio_technique and esi fund exceptions are here because inconsistencies between the models and schema
make them hard to wrangle into a general test.