Replace multiple select with Add Another
Add another component has been added to the new document form. The component has a lot of spacing around it which means it doesn't sit well within the component. We are removing the additional styling we've added to the component when we previously added them to the change request form. This will put the styling for the add another component consistent with the Whitehall fatality notice Add Another component.

They were also hooked into a .js class and were overriding a gem component, both of which are discouraged practices. Gem components are built aorund a concept of 'component isolation' which precludes this sort of ad-hoc CSS override. Any styling of a publishing gem component should be done in the gem, not within the apps.