Switch legacy and new design system view naming
This is to make it easier to cherry-pick commits since the commits that transitions the view to the new design system is done on a file named 'new'. There would be a lot of effort to move the changes between files if we kept the convention as `new` and `new_design_system`. Switching to `new_legacy` and `new`