Do not rendering existing error messages for design system
The method for generating document_error_messages modifies the error messages and make them html safe, which breaks the new layout in the design system.
Before this change, the layout markup breaks after trying to save the form.
The change keeps for layout intact. we will need to think about how we do error message validation for design system at a later point.