Ported over govspeak editor JS and CSS from Whitehall
Copied over files for javascript and styling. Removed some functionality that are currently not used in Specialist Publisher.
e.g. Rendering of Images and Attachment within the GovSpeak Editor preview. This is because Specialist Publisher does not support image uploading currently, and attachments are only available on the edit screens, which are not yet transitioning to the new design system. Removing the unused code instead of just leaving them there because the way it is implemented is different in Whitehall, e.g. the way we gather the image IDs and Attachment IDs. We can probably make that decision when we get to the point of transitioning Attachments and hence requiring to preview them. Another card has been created for Preview Button for Attachments.