Switch unsaved changes spec to use webdriver BiDi protocol
The WebDriver spec requires drivers to ignore browser dialogs and Chrome's web driver is faithful to this since Chrome version 127. See https://issues.chromium.org/issues/42323796 for more info.

This stops our unsaved changes spec passing because they need to interact with the confirmation dialog when the user navigates away from a page with unsaved changes.

However, drivers using the BiDi protocol are able to interact with browser dialogs. We therefore switch the unsaved changes spec to use a driver with the BiDi protocol enabled. We do not enable the protocol for all specs because Selenium's BiDi implementation is not as feature complete as its default driver and it also seems to be quite slow.