Remove broken tests
This was disabled in https://github.com/alphagov/specialist-publisher/pull/2715
when the tests started to fail due to a new release of Chrome on
CI.

Despite extensive attempts to re-enable the tests, we've so far
been unsuccessful:
https://trello.com/c/7FTvIbPD

The value of the tests is not worth the effort needed to get them
working again. Moreover, these tests will be made obsolete by our
moving to the Design System, which is already in progress.
Therefore we've decided to delete the tests ahead of time.