Add validation for multiple select
In order to add validation, we need to show the item of index 0.
In adding a default initial element upon render, the delete button does not work for the first element (consistent with the behaviour we're seeing where if the form re-renders, pre-existing add another items are non-deleteable. We will be moving to use select with search in the near future so this will go away soon.