Delete the test folder that is not running any tests
The autocomplete test has never been run since it lives in the test directory and specialist publisher is Rspec.
The test will also possibly not run since helper methods and inherited base class neither of which exist in the Specialist Publisher repo.
Autocomplete also does not exist as a component, but is a partial with some javascripts, used in the admin views. We can look at alternative testing methods later.
Deleting this directory for now to not be confusing for people, and to stop further tests from being added to a directory that doesn't get run.