Extend access to finders index to non GDS users
This page should be available to all users. Users will only see it as the root page of the app if they can preview the design system.

As part of this commit, we tidy up the finder schema policy and its usage, so that we're making better use of the conventions provided by the Pundit gem we use for authorisation