Create view component for finders index page
Scoping the list of finders required some refactoring of the application controller methods so that the authorisation logic can be shared effectively for both documents and finder schemas

The design originally called for rendering the finders in a table and including information about the publishing state and last update time for the finder. However, this would have required us to fetch that information from Publishing API. The additional complexity of fetching the data from Publishing API was deemed not worth the effort, so we have switched to using the document list component to render the finders list, as we only need to render the name of the finder and a link to its summary page.