Skip authorisation for finders index action
There's no need to authorize the finders index action, as all users should be able to view the list of finders scoped to their permission level. The index method on the finder schmea policy was failing for non GDS editor users because it was attempting to check the schema organisations for the finder schema class, when no such method exists.

To avoid reoccurrence of such issues in future I have moved any behaviour in the application policy which was in fact specific to documents into the document policy class.