Use 'document_type' as convention for naming views
Instead of using the `document_title` for naming views, use the `document_type` instead. On occasion, we may want to change the title of a finder (and therefore documents). We want to be able to show this new name across the publisher app. Nonetheless, we should not be required to change the already configured models/views/schema file names.

This change ensures that all the view names now match those of our models. Title changes will not affect existing file and object names.