Use 'document_type' as convention for naming views
Instead of using the `document_title` for naming views and loading partials, use the `document_type` instead. On occasion, we may want to change the title of a finder (and therefore documents). We want to be able to show this new name across the publisher app. Nonetheless, we should not be required to change the already configured models/views/schema file names.

Previously, the views were named and loaded based on the `document_type_slug`, passed as a parameter with the value of `admin_slug` (`title.parameterize.pluralize`), which is title based.

This change ensures that all the view names now match those of our models. Title changes will not affect existing file and object names.

With this change, `document_type` (which is defined as the document model's `to_s.underscore`) dictates both the internal naming conventions within SP, and the actual document type we send to Publishing API (as per the `DocumentPresenter`'s `document_type: document.document_type` field). This means we cannot rename models within SP without causing issues downstream. This will be addressed in a future commit.