Include additional document types in generic testing
 Document types `marine_equipment_approved_recommendation` and
`protected_food_drink_name` were not included because they were failing validations due to custom year and time fields.

Added a tiny bit of logic to deal with those cases. This makes it more obvious that we're making use of such date variations (for future design system date component improvements), and it enables us to run the entire test suite for these document types, so we don't have gaps in the coverage.