Allow specifying `document_type` for downstream usage, in finder schema
These changes introduce a new schema field, "downstream_document_type", which will be used at the `DocumentPresenter` level to send the document type to Publishing API and in other cases where we talk to Publishing API. Historically, what we end up sending to Publishing API as the `document_type` is defined in the `Document` model as `to_s.underscore`, meaning it's tied to the name of that specific model. This typically matches the value of the `filter.format`, causing a general belief that this schema define format is the document type. We can see in ESIFund's case, that we specify its metadata in Publishing API as `european_structural_investment_fund_metadata` (in `_specialist_document.jsonnet`), but the "allowed" document_type is set as `esi_fund`. In fact, `filter.format` is only relevant later on, for Search API.

Whilst we can enforce the `document_type` to be the same as `filter.format` as a convention (will add guidance in later commit), we already have cases that do not follow the convention (ESI Fund). It is difficult to alter these without migrating the existing documents.

The `DocumentFinder` will be using the `document_type` from the Publishing API payload to find the relevant model, so we will need to maintain a map of the diverging document types. See next commit that implements `EsiFund` changes.