Introduce a mapper for non-conforming document types
This commit changes esi funds finder related files to follow the generic naming conventions we have elsewhere: the `filter.format` field in the schema is used to name json, model and views. This is not enforced, but it is helpful when navigating the codebase. Additionally, it sets the document type Publishing API expects for EsiFunds by using the new `downstream_document_type` field. This is to make it clear that the type is not derived from the `filter.format`, but from the model name (or downstream schema value, if set).

It is desirable to be able to change the name of files in this repo in this manner, without breaking anything downstream. Nonetheless, the document type from the Publishing API payload is used to look up the models in the `DocumentFinder`. A mapper helper file is now used to store any outliers where the `downstream_document_type` differs from how we name models/schemas/views.

Only ESI funds is noted as an outlier, and future commit will add guidance on naming conventions.

Having the `downstream_document_type` field visible in the schema and presenter makes it clearer how exactly the document type is set for Publishing API. The model's `document_type` field is now used for internal SP concerns, and still sets the document type sent to Publishing API, but is aware of defined mappings where conventions don't apply. The `filter.format` is mainly used for Search API purposes.