Include additional document type in generic testing
Document type `marine_equipment_approved_recommendation` was excluded from generic testing because it was failing validations due to a custom text field that takes a year value.

Added a tiny bit of logic to deal with this case. This makes it more obvious that we're making use of such date variations (for future design system date component improvements), and it enables us to run the entire test suite for this document types, so we don't have gaps in coverage.

`protected_food_drink_name` should have also been included in generic testing (it has a text field that takes a validated time value, similar to the above case). Nonetheless, for some reason it passes validation with no summary, thus breaking some tests. I tested in deployed environments too, it seems to be the case everywhere. Could not get to the bottom of it, so left it as an exception for now.