Migrate business finance support scheme
Testing:
- model is covered in `document_type_spec.rb` so no need to have its own model test (now deleted)
- feature test coverage: added itw own test file as it has custom fields. This can be refactored at a later point.

View:
- use new `InputComponent` to render custom `:continuation_link` and `:will_continue_on` fields

Had to reorder the model validations so that the errors are show in "visual" top-down order in the ErrorSummary, matching the order in which the fields are displayed.

Note for future:
The two custom fields are actually sent to Publishing API as metadata and then custom front end rules apply for rendering a big green continue button based on these values. They can easily be sent as non-filterable facets, and then we can include them in the schema, with custom validations (the link is url-validated). A similar approach was taken for Licence Transaction. Alternatively, they can be included in general testing as known, but somewhat standard, exceptions - standard in the sense that there might be more than one document type using them.