Split attachment testing into their own files
Currently the only attachment tests exist for cma_cases. Since most of them share the same attachments functionality, it should be sufficient. We may want to move it to generic testing to run for multiple document types but need to consider the time we'd be adding to tests and the value that brings in running attachments tests multiple times for diffefent document types.

Also split them into design system spec and legacy spec to prepare for transition.