Refactor design system views
We want to only ever have to define the custom views for non-conforming pages. The standard pages just render the `specialist_document_form` partial. This makes it more obvious which documents have custom views.

Using a lookup_context implementation in a controller helper method. This looks up for a defined partial, and falls back on `specialist_document_form` if it cannot find the partial.