Refactor legacy views
We want to only ever have to define the custom views for non-conforming pages. The standard pages just render the `specialist_document_form` partial. This makes it more obvious which documents have custom views.

Using a lookup_context implementation in a controller helper method. This looks up for a defined partial, and falls back on `specialist_document_form` if it cannot find the partial.

Adding a legacy method as to distinguish it from the DS helper method. This method should be deleted alongside the rest of the code when the DS is released.

The edit view has not yet been migrated to the Design System so it's still using the legacy partials.