Add confirmation page for publishing flow
Publishing flow now means users are taken to an interstitial page when they click publish. The publish button is displayed conditionally in a previous PR for the document summary page.
In that previous PR, publishing text was displayed in a notice box under the actions button. Now that we have a confirmation page, we've moved it there.
Created a new set of publishing text since we are only going to ever see this text when the Publishing button is available and takes you to the Publish page.
So it only needs to show publishing text in the cases where:
- there is a first draft
- there is a published with major edit draft
- there is a published with minor edit draft
made the old publishing text legacy, since the other cases in the legacy publishing text method deals with the case where Publishing button does not show. These are not particularly useful, and the permissions information can be displayed in the notice box. This will be dealt with in another PR.