Add confirmation page for publishing flow.
Updated presenter publishing text to display what is in the designs.
Remove the publishing text on the show page.
Created a new set of publishing text since we are only going to ever see this text when the Publishing button is available and takes you to the Publish page.
So it only needs to show text in the cases where:
- there is a first draft
- there is a published with major edit draft
- there is a published with minor edit draft
made the old publishing text legacy