Reintroduce single select without search for legitimate use-cases
There are a few fields on the con-conforming pages that have a small number of options, and it makes sense for them to be preselected. This is still a valid case for single select without search. This also preserves the existing behaviour for these fields, in line with our decision to not roll out any changes to the UI where we can avoid it.

Since this component is now only used for this type of select with no blank option, there is no need to have that parameter anymore. It can be reintroduced in the future if there is a need for it.

In the future, if the non-conforming documents get "generalised", schemas could include another specification in the "specialist publisher properties" field, whether a single select (or multiselect) facet should have the "with search" feature.