Add tests for design system publish flow
The feature tests are doing too much, moving a lot of the show/hide button testing into the sidebar component tests. Feature test should test the actual flow of publishing. There are still some integration type tests in there that is testing whether publishing API and asset manager are called, which should also be moved, but can probably happen later as it's out of scope, we can move them  next time we touch the integrations.