Selects with a multiple attribute should show several options (no JS)
Select elements are set to 2.5rem high, which is a defined within
`govuk-frontend`. This has a detrimental affect on selects using the
`multiple` attribute as they only display one option. It's even worse
if the first option is blank as the select appears to be empty, and
it's easy to miss the tiny scroll bar that reveals more options.

Setting the height of these variant selects to `auto` will ensure that
they display several options within the scrolling area and therefore
will provide an improved UX.

Note that in practice this only applies to the non-JS experience:
selects with a `multiple` attribute are entirely replaced in the JS
experience with markup generated by the third party plugin choices.js.

Consideration:

This fix should apply to all apps that use `select_with_search` with
multiple variants. Therefore it should be made available in a higher
root dependency, for example the publishing gem or the design system.