Show only relevant unpublishing warning messages
In the old world, there used to be a permanent "Unpublishing" section which used to show a message of why the unpublish button wasn't showing or what the impact would be of unpublishing.
Now that we have an interstital page, we are duplicating the information when the button is visible. Removing the need to show warning message on summary page as it already shows on the confirmation page.
Also removing some other scenarios and instances of the notice messages appearing such as: when document is already unpublished or when it is first draft. These can be seen on the publish status of the document. The only relevant useful message we should show users is when they expect to see an unpublishing button but it's not there. This is the case where they have a publish with draft, or they don't have permission. these messages will still be showing.