Move all transitioned feature specs to legacy folder
So we know which feature specs have already been considered and an equivalent design system one has been made.
This will help in the cleanup phase where we just just straight up delete the entire folder. The feature specs outside of these folders either have not yet been transitioned to design system or have not been considered and need to be reviewed.