Switched edit page design system to use the new shared templates with :new
A lot of the partials are shared with the new page. With the exception of the minor major changes and attachment links which will be taken care of in the next commit.
Changed the Save button text to "Save" to be consistent with Figma designs.
Skipped some attachment flow tests as they were failing for Publishing API integration expectations, which were stubbed within the tests. We will be revisiting the attachment flow in a later story so should rewrite these tests at that point to match the new flow.