Transition the Attachment links area to new design system
The delete attachment button does nothing at the moment. This is because in the legacy world, this was a POST button that discards the attachment. In the Design system world, this should take people to a confirm delete attachment page that will be introduced as part of a later story.