Date component fields should load values from existing documents
Date component fields were not pre populating existing values, so editing existing drafts result in an empty date field even when there is a value.
Added tests for the view component. Pulled Params out as injected dependency for ease of testing.
There's two scenarios where we want to pre populate the date values, on first edit where the values are loaded from the document model. Or on form submit reload where the date values come from params. In the case where the form is submitted and reloaded, the model is also updated with the form values. In these cases, it is possible to have document models with invalid date values, such as "not-a-date".
This is why there's logic introduced in value pre-poluating for date fields where, if there are any values populated in params, use that instead, and if there's nothing from params then load it from the document model. In the case where we want to load it from document model, it is only a string, so in order to parse it into 3 parts, we need to parse it as a date first.