[Bug fix] Design system drop down should work for nested facets
Use FacetSelectHelper in the 3 facet select inputs to render correct nested facet labels.
Moved method in Facet Select Helper out of the private section since it is used in the admin pages views.
Tests wise, for design system need to split testing for creating new trademark decision from the more generic tests since it contains nested facets for dropdowns which require us to select the concatenated version of parent and child facet value. The generic test does not support this as it attempts to select the first allowed value, which does not exist in a nested facet dropdown. Unsure how this is working for legacy tests.... but manual test for legacy works and the dropdown only lists the concatenated version. Not sure how the generic tests in legacy is working for trademark decisions as it is not handling nested facets any specially.