Get Generating Preview message for Govspeak Editor preview to display correctly
The call to generate preview was previously a synchronous call. This freezed the process and nothing was being rendered on the UI until the call completed. Changing this to Async, and also got the generated content to be wiped and replaced with a "Generating..." message each time the user clicks Preview.