Add tests for design system attachments
The previous tests for attachments are convoluted and are not actually testing the behaviours such as adding, editing, and deleting attachments.
This commit separates the existing editing attachments test and attachments validation tests into 2 test files: attachments test where we test adding, editing and deleting attachments along with asset manager interaction. And a govspeak test where we test govspeak validation and govspeak preview behaviours

The inline attachments validation does not need to escape HTML for the design system, but it needs to do so for the legacy view. Moved this to controller level so we can better control what happens in legacy view. This can also be subsequently removed on cleanup.