Make breadcrumbs consistent across all pages
Breadcrumbs should all contain a link to the All finders page as the root.
Breadcrumbs should also say "<format plural> finder" if there is a link back to the finder page to match the design on Figma.
Had to fix one test which was stubbing out all methods of a Document model. Previously title was returning nil.