Revert changes to breadcrumbs for document pages
The changes in the previous commit were made to make the breadcrumbs consistent with the index page title which was implemented as `"#{current_format.title} finder"` in the Design System.

Whilst that was good for consistency, I believe it's better to preserve the pre-design-system naming conventions - the title of the documents index page was `current_format.title.pluralize` which is more in line with the function of the page, i.e. creating new documents (of that format). Additionally, the links on the "All finders" page, that take us to the "finder" documents index page, are given as `current_format.title.pluralize` (without the "finder" bit), so it would be confusing to the user to navigate to a page whose title doesn't match the link they clicked on.

The finder setup admin area can continue using the "finder" in the title, since those pages actually deliver finder-related functionality.